Multiple audio track support (#76)
* Add support for two VirtualSourceBuffers (one audio and one video/combined)
Add activeSourceBuffers property to HtmlMediaSource and update on each buffer's first updateend, as well as on the player's audioTracks change event. Create independent transmuxers for each source buffer.

* Move the real source buffers up one level to the "virtual" mediaSource so that real sourceBuffers can be easily shared among the virtualsourcebuffers

* Fix watch tasks in package.json
Now watch for changes in es6-ified depenencies and rebuild as necessary

* support for audio/video only and combined streams

* Don't consider the `audioBuffer` when determining the VirtualSourceBuffer's `updating` state if `audioDisabled_` is not false

* doc comments for the whole project
added @file to the top of all files
some minor style updates
changed URL/MediaSource to classes with static elements
added doc watching to package.json
remove eslint-disable around max-len urls due to vjsstandard update

* Always return a boolean from the VirtualSourceBuffer's updating getter

* Use mux.js builds with support for AAC audio

* emit info events when tracks change (#82)

If interesting video or audio features change, emit an event that can be reacted to downstream. For instance, changing the audio sample rate may be a problem for some decoders (e.g. Firefox), and this event can alert you to prevent a decode error.

* Remove global videojs, prefer import videos module. (#72)

* Add isTypeSupported method existing checker (#83)

if MediaSource object exists but isTypeSupported method doesn't exist, script and videoJS will be broken.

This issue can be reproduced in Opera TV OS or Philips Android TV which uses Opera as built-in HTML5 application browser.

* Adding support for video and audio only playlists (#85)

Defer the creation of the underlying real SourceBuffers until one of two conditions is fulfilled: 1) we get data back from the transmuxer so that we know the track layout of the source 2) a second VirtualSourceBuffer is created meaning that both video and audio source buffers are necessary
* Renamed the `mediaSource_` property to `nativeMediaSource_` for clarity

* Fixed a bug where we weren't forcing the creation of real sourcebuffers on the second virtual sourcebuffer (#86)

* Round the buffered values to compensate for Flash's noisey values (#87)

The vjs_buffered property on the SWF can annoyingly returns very slightly different values between appends that appears to come from type conversion between Flash and the JavaScript engine. This can confuse our segment annotation since it thinks the values have changed when if they havent and adds the wrong end-time information into the segment as a result

* Update the version of mux.js to use 2.2.0 - the latest non-prerelease with multiple alternate audio track support